Reusable service-level get operations for companies
